Make Users.Manager gen server compatible with async tests
Instead of starting a global Users.Manager process, support
starting a separate one for each test process when Mix.env() == :test.

This makes both the implementation and the test code somewhat noisier.
But the upside is that we get full control of the manager process in
tests and obtain the ability to mock it independently in individual
test cases.